tests/Grid_API_Testing/requirements.txt
termcolor
requests
paramiko
configparser
nose-testconfig
nose
git+https://github.com/gigforks/packet-python.git
git+https://github.com/zero-os/0-orchestrator.git#subdirectory=pyclient
git+https://github.com/zero-os/0-core.git#subdirectory=client/py-client
